optional rich text pdf export
