+++ backend/tests_configs.ini
@@ -315,6 +315,7 @@ email.notification.enabled_on_invitation = False
 webdav.ui.enabled = False
 webdav.base_url = https://localhost:3030
 webdav.root_path = /webdav
+rich_text_preview.enabled = True
 
 [functional_translation_test]
 app.enabled = contents/thread,contents/file,contents/html-document,contents/folder,upload_permission,share_content
