+++ backend/tracim_backend/models/context_models.py
@@ -84,6 +84,7 @@ class ConfigModel(object):
         limitation__maximum_online_users_message: str,
         call__enabled: bool,
         call__unanswered_timeout: int,
+        rich_text_preview__enabled: bool,
     ) -> None:
         self.email_notification_activated = email_notification_activated
         self.new_user_invitation_do_notify = new_user_invitation_do_notify
@@ -105,6 +106,7 @@ class ConfigModel(object):
         self.limitation__maximum_online_users_message = limitation__maximum_online_users_message
         self.call__enabled = call__enabled
         self.call__unanswered_timeout = call__unanswered_timeout
+        self.rich_text_preview__enabled = rich_text_preview__enabled
 
 
 class ErrorCodeModel(object):
